+++ Snyk.VisualStudio.Extension.2022/Settings/SnykGeneralSettingsUserControl.cs
@@ -8,6 +8,7 @@ using System.Windows.Forms;
 using Microsoft.VisualStudio.Shell;
 using Microsoft.VisualStudio.Threading;
 using Serilog;
+using Snyk.VisualStudio.Extension;
 using Snyk.VisualStudio.Extension.Authentication;
 using Snyk.VisualStudio.Extension.CLI;
 using Snyk.VisualStudio.Extension.Extension;
@@ -92,7 +93,7 @@ namespace Snyk.VisualStudio.Extension.Settings
         {
             this.authenticateButton.Enabled = LanguageClientHelper.IsLanguageServerReady();
             this.customEndpointTextBox.Text = snykOptions.CustomEndpoint;
-            this.organizationTextBox.Text = snykOptions.Organization;
+            this.organizationTextBox.Text = string.IsNullOrEmpty(snykOptions.Organization) ? "Configure in Solution Settings" : snykOptions.Organization;
             this.ignoreUnknownCACheckBox.Checked = snykOptions.IgnoreUnknownCA;
             this.tokenTextBox.Text = snykOptions.ApiToken.ToString();
 
@@ -265,11 +266,6 @@ namespace Snyk.VisualStudio.Extension.Settings
             this.errorProvider.SetError(this.customEndpointTextBox, "Needs to be a full absolute well-formed URL (including protocol)");
         }
        
-        private void OrganizationInfoLink_LinkClicked(object sender, LinkLabelLinkClickedEventArgs e)
-        {
-            this.OrganizationInfoLink.LinkVisited = true;
-            Process.Start("https://docs.snyk.io/ide-tools/visual-studio-extension#organization-setting");
-        }
 
         private void authType_SelectionChangeCommitted(object sender, EventArgs e)
         {
@@ -278,10 +274,6 @@ namespace Snyk.VisualStudio.Extension.Settings
                 LanguageClientHelper.LanguageClientManager().DidChangeConfigurationAsync(SnykVSPackage.Instance.DisposalToken).FireAndForget();
         }
 
-        private void organizationTextBox_TextChanged(object sender, EventArgs e)
-        {
-            snykOptions.Organization = organizationTextBox.Text;
-        }
 
         public Panel GetPanel()
         {