+++ Snyk.VisualStudio.Extension.2022/Settings/SnykSolutionOptionsUserControl.cs
@@ -2,6 +2,7 @@
 using System.Windows.Forms;
 using Microsoft.VisualStudio.Shell;
 using Serilog;
+using Snyk.VisualStudio.Extension;
 using Snyk.VisualStudio.Extension.Service;
 
 namespace Snyk.VisualStudio.Extension.Settings
@@ -15,6 +16,7 @@ namespace Snyk.VisualStudio.Extension.Settings
 
         private readonly ISnykServiceProvider serviceProvider;
         public string AdditionalOptions { get; set; }
+        public string Organization { get; set; }
         /// <summary>
         /// Initializes a new instance of the <see cref="SnykSolutionOptionsUserControl"/> class.
         /// </summary>
@@ -29,6 +31,27 @@ namespace Snyk.VisualStudio.Extension.Settings
         private void AdditionalOptionsTextBox_TextChanged(object sender, EventArgs e)
         {
             AdditionalOptions = this.additionalOptionsTextBox.Text;
+            // Notify Language Server of configuration change
+            ThreadHelper.JoinableTaskFactory.RunAsync(async () =>
+            {
+                await this.serviceProvider.LanguageClientManager.DidChangeConfigurationAsync(SnykVSPackage.Instance.DisposalToken);
+            }).FireAndForget();
+        }
+
+        private void OrganizationTextBox_TextChanged(object sender, EventArgs e)
+        {
+            Organization = this.organizationTextBox.Text;
+            // Notify Language Server of configuration change
+            ThreadHelper.JoinableTaskFactory.RunAsync(async () =>
+            {
+                await this.serviceProvider.LanguageClientManager.DidChangeConfigurationAsync(SnykVSPackage.Instance.DisposalToken);
+            }).FireAndForget();
+        }
+
+        private void OrganizationInfoLink_LinkClicked(object sender, LinkLabelLinkClickedEventArgs e)
+        {
+            this.OrganizationInfoLink.LinkVisited = true;
+            System.Diagnostics.Process.Start("https://docs.snyk.io/ide-tools/visual-studio-extension#organization-setting");
         }
 
         private void SnykProjectOptionsUserControl_Load(object sender, EventArgs eventArgs) => ThreadHelper.JoinableTaskFactory.RunAsync(async () =>
@@ -36,6 +59,7 @@ namespace Snyk.VisualStudio.Extension.Settings
             try
             {
                 var additionalOptions = await this.serviceProvider.SnykOptionsManager.GetAdditionalOptionsAsync();
+                var organization = await this.serviceProvider.SnykOptionsManager.GetOrganizationAsync();
 
                 if (!string.IsNullOrEmpty(additionalOptions))
                 {
@@ -45,12 +69,22 @@ namespace Snyk.VisualStudio.Extension.Settings
                 {
                     this.additionalOptionsTextBox.Text = string.Empty;
                 }
+
+                if (!string.IsNullOrEmpty(organization))
+                {
+                    this.organizationTextBox.Text = organization;
+                }
+                else
+                {
+                    this.organizationTextBox.Text = string.Empty;
+                }
             }
             catch (Exception e)
             {
-                Logger.Error(e, "Error on load additional options");
+                Logger.Error(e, "Error on load additional options and organization");
 
                 this.additionalOptionsTextBox.Text = string.Empty;
+                this.organizationTextBox.Text = string.Empty;
             }
         });
     }