+++ Snyk.VisualStudio.Extension.2022/Settings/SnykSolutionSettings.cs
@@ -19,6 +19,11 @@
         public string AdditionalOptions { get; set; }
 
         /// <summary>
+        /// Gets or sets organization for project.
+        /// </summary>
+        public string Organization { get; set; }
+
+        /// <summary>
         /// Gets or sets a value indicating whether Is all projects scan enabled.
         /// </summary>
         public bool IsAllProjectsScanEnabled { get; set; }